Testimonial design changed
