+++ frontend/src/components/Testimonial.jsx
@@ -1,3 +1,4 @@
+
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { Pagination, Navigation, Autoplay } from 'swiper/modules';
 import 'swiper/css';
@@ -5,21 +6,16 @@ import 'swiper/css/pagination';
 import 'swiper/css/navigation';
 import '../index.css';
 
-import Polygon1 from "./Testimonials/Polygon1";
-import Polygon2 from "./Testimonials/Polygon2";
-import Polygon3 from "./Testimonials/Polygon3";
-import Polygon4 from "./Testimonials/Polygon4";
-import Polygon5 from "./Testimonials/Polygon5";
-import Polygon6 from "./Testimonials/Polygon6";
+import Polygon from "./Testimonials/Polygon";
 
 const Testimonial = () => {
   return (
-    <div className="min-h-screen px-0 md:px-4 py-16 pb-10 md:pb-32 text-center relative overflow-visible">
+    <div className="px-0 md:px-8 py-16 pb-10 md:pb-32 my-20 text-center relative overflow-visible">
 
       <h1 className="text-4xl sm:text-5xl font-bold text-[#002639] mb-2">
         “Voices of Success”
       </h1>
-      <p className="text-[#002639] font-medium mb-20 text-2xl sm:text-3xl">
+      <p className="text-[#002639] font-medium mb-5 text-2xl sm:text-3xl">
         Our Student Stories
       </p>
 
@@ -33,22 +29,25 @@ const Testimonial = () => {
         autoplay={{ delay: 5000, disableOnInteraction: false }}
         breakpoints={{
           768: {
-            slidesPerView: 3,
-            slidesPerGroup: 3,
+            slidesPerView: 2,
+            slidesPerGroup: 2,
           },
           1024: {
             slidesPerView: 3,
             slidesPerGroup: 3,
           },
         }}
-        className="max-w-full mx-auto relative pb-20"
+        className="max-w-screen-xl mx-auto relative pb-5"
       >
-        <SwiperSlide><Polygon1 /></SwiperSlide>
-        <SwiperSlide><Polygon2 /></SwiperSlide>
-        <SwiperSlide><Polygon3 /></SwiperSlide>
-        <SwiperSlide><Polygon4 /></SwiperSlide>
-        <SwiperSlide><Polygon5 /></SwiperSlide>
-        <SwiperSlide><Polygon6 /></SwiperSlide>
+        <SwiperSlide><Polygon /></SwiperSlide>
+        <SwiperSlide><Polygon /></SwiperSlide>
+        <SwiperSlide><Polygon /></SwiperSlide>
+        <SwiperSlide><Polygon /></SwiperSlide>
+        <SwiperSlide><Polygon /></SwiperSlide>
+        <SwiperSlide><Polygon /></SwiperSlide>
+        <SwiperSlide><Polygon /></SwiperSlide>
+        <SwiperSlide><Polygon /></SwiperSlide>
+        <SwiperSlide><Polygon /></SwiperSlide>
       </Swiper>
     </div>
   );
