+++ frontend/src/index.css
@@ -38,218 +38,6 @@ html, body {
     @apply hidden
 }
 
-/* Testimonial shapes */
-
-.shape-1 {
-  width: 270px;
-  height: 290px;
-  border: 1px solid black;
-  border-radius: 15%;
-  background-color: white;
-  transform: skew(-5deg, 20deg);
-  position: relative;
-  z-index: 2;
-}
-
-.shadow-shape-1 {
-  width: 270px;
-  height: 290px;
-  border-radius: 15%;
-  background-color: #2d7b67;
-  position: absolute;
-  transform: skew(-10deg, 25deg);
-  top: 20px;
-  left: -15px;
-  z-index: 1;
-}
-
-.shape-2 {
-  width: 350px;
-  height: 450px;
-  border: 1px solid black;
-  border-radius: 15%;
-  background-color: white;
-  transform: skewY(-5deg) perspective(800px) rotateX(30deg);
-  position: relative;
-  z-index: 2;
-}
-
-.shadow-shape-2 {
-  width: 350px;
-  height: 400px;
-  border-radius: 15%;
-  background-color: #2d7b67;
-  position: absolute;
-  transform: skewX(5deg) perspective(800px) rotateX(30deg);
-  top: -1px;
-  left: 10px;
-  z-index: 1;
-}
-
-.shape-3 {
-  width: 270px;
-  height: 290px;
-  border: 1px solid black;
-  border-radius: 15%;
-  background-color: white;
-  transform: skew(5deg, -20deg);
-  position: relative;
-  z-index: 2;
-}
-
-.shadow-shape-3 {
-  width: 270px;
-  height: 290px;
-  border-radius: 15%;
-  background-color: #2d7b67;
-  position: absolute;
-  transform: skew(10deg, -25deg);
-  top: 20px;
-  left: 15px;
-  z-index: 1;
-}
-
-@media (max-width: 1024px) {
-  .shape-1 {
-      width: 230px;
-      height: 250px;
-      border: 1px solid black;
-      border-radius: 15%;
-      background-color: white;
-      transform: skew(-5deg, 20deg);
-      position: relative;
-      z-index: 2;
-    }
-  
-    .shadow-shape-1 {
-      width: 230px;
-      height: 250px;
-      border-radius: 15%;
-      background-color: #2d7b67;
-      position: absolute;
-      transform: skew(-10deg, 25deg);
-      top: 20px;
-      left: -15px;
-      z-index: 1;
-    }
-  
-    .shape-2 {
-      width: 320px;
-      height: 420px;
-      border: 1px solid black;
-      border-radius: 15%;
-      background-color: white;
-      transform: skewY(-5deg) perspective(800px) rotateX(30deg);
-      position: relative;
-      z-index: 2;
-    }
-  
-    .shadow-shape-2 {
-      width: 320px;
-      height: 370px;
-      border-radius: 15%;
-      background-color: #2d7b67;
-      position: absolute;
-      transform: skewX(5deg) perspective(800px) rotateX(30deg);
-      top: -1px;
-      left: 10px;
-      z-index: 1;
-    }
-  
-    .shape-3 {
-      width: 230px;
-      height: 250px;
-      border: 1px solid black;
-      border-radius: 15%;
-      background-color: white;
-      transform: skew(5deg, -20deg);
-      position: relative;
-      z-index: 2;
-    }
-  
-    .shadow-shape-3 {
-      width: 230px;
-      height: 250px;
-      border-radius: 15%;
-      background-color: #2d7b67;
-      position: absolute;
-      transform: skew(10deg, -25deg);
-      top: 20px;
-      left: 15px;
-      z-index: 1;
-    }
-}
-
-@media (max-width: 768px) {
-  .shape-1 {
-    width: 300px;
-    height: 320px;
-    border: 1px solid black;
-    border-radius: 15%;
-    background-color: white;
-    transform: skew(-5deg, 20deg);
-    position: relative;
-    z-index: 2;
-  }
-
-  .shadow-shape-1 {
-    width: 300px;
-    height: 320px;
-    border-radius: 15%;
-    background-color: #2d7b67;
-    position: absolute;
-    transform: skew(-10deg, 25deg);
-    top: 20px;
-    left: -15px;
-    z-index: 1;
-  }
-
-  .shape-2 {
-    width: 350px;
-    height: 450px;
-    border: 1px solid black;
-    border-radius: 15%;
-    background-color: white;
-    transform: skewY(-5deg) perspective(800px) rotateX(30deg);
-    position: relative;
-    z-index: 2;
-  }
-
-  .shadow-shape-2 {
-    width: 350px;
-    height: 400px;
-    border-radius: 15%;
-    background-color: #2d7b67;
-    position: absolute;
-    transform: skewX(5deg) perspective(800px) rotateX(30deg);
-    top: -1px;
-    left: 10px;
-    z-index: 1;
-  }
-
-  .shape-3 {
-    width: 300px;
-    height: 320px;
-    border: 1px solid black;
-    border-radius: 15%;
-    background-color: white;
-    transform: skew(5deg, -20deg);
-    position: relative;
-    z-index: 2;
-  }
-
-  .shadow-shape-3 {
-    width: 300px;
-    height: 320px;
-    border-radius: 15%;
-    background-color: #2d7b67;
-    position: absolute;
-    transform: skew(10deg, -25deg);
-    top: 20px;
-    left: 15px;
-    z-index: 1;
-  }
-}
 
 /* Swiper */
 
@@ -258,7 +46,7 @@ html, body {
   color: #2d7a66 !important;
   width: 30px;
   height: 30px;
-  top: 45%;
+  top: 55% !important;
   transform: translateY(-50%);
   z-index: 10;
 }
@@ -268,8 +56,15 @@ html, body {
   color:#204b54;
 }
 
-/* Hide arrows on mobile (screens smaller than 768px) */
-@media (max-width: 1024px) {
+.swiper-button-prev {
+  left: -30px !important;
+}
+
+.swiper-button-next {
+  right: -30px !important;
+}
+
+@media (max-width: 1023px) {
   .swiper-button-prev,
   .swiper-button-next {
     display: none !important;
@@ -281,16 +76,16 @@ html, body {
 }
 
 .swiper-pagination {
-  bottom: -70px !important;
+  bottom: -30px !important;
   position: absolute !important;
   width: 100%;
   text-align: center;
 }
 
 .swiper-pagination-bullet {
-  background-color: #737776 !important;
-  width: 18px;
-  height: 18px;
+  background-color: #5c5e5e !important;
+  width: 24px;
+  height: 24px;
   border-radius: 50%;
   opacity: 1;
   margin: 0 6px !important;
@@ -300,6 +95,6 @@ html, body {
   background-color: #2d7a66 !important;
   width: 26px;
   height: 26px;
-  transform: scale(2.0);
+  transform: scale(2.5);
   transition: all 0.3s ease;
 }
